+++ CHANGELOG.md
@@ -6,4 +6,4 @@
 * Add @Redblack tree definition and functions
 * Get rid Stack in favor new Nix-style Cabal
 * Add @degree in @Graph strucutre
-* Add @Injected and @Embedded type families
+* Add @Injected, @Embedded and @Attached type families